LIR parser
